chore: rework headers introspection with headers builder chore: split index.js into smaller subject files
